+++ migrations/2019-02-12-202203_central_repo_db/down.sql
@@ -2,6 +2,7 @@ DROP TABLE Action;
 DROP TABLE Category;
 DROP TABLE Qualifier;
 DROP TABLE Priority;
+DROP TABLE RawEvent;
 DROP TABLE Status;
 DROP TABLE Clusters;
 DROP TABLE Outliers;