+++ CHANGELOG.md
@@ -5,6 +5,8 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added a new table called `RawEvents` in database schema.
 
 ## [0.5.0] - 2019-07-24
 ### Added